// todo-ui/src/app/services/models/page-response-task-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TaskResponse } from '../models/task-response';
export interface PageResponseTaskResponse {
  content?: Array<TaskResponse>;
  isFirst?: boolean;
  isLast?: boolean;
  page?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}